Django Setup and Response
You can start the server by typing in "python manage.py runserver" in the main dir.

After that you can call "http://127.0.0.1:8000/send_qr/<AnyID>/" to send a request to the server.
Right now the server will only respond with a 200 with the ID "nas80dtw3ß28zasd0na0", else it will throw a 404.